+++ tests/TestCase/ScanEntityTest.php
@@ -35,7 +35,7 @@ class ScanEntityTest extends TestCase
     {
         parent::setUp();
 
-        $this->ScanEntity = new ScanEntity([
+        $this->ScanEntity ??= new ScanEntity([
             'code' => 200,
             'external' => false,
             'location' => 'https://example.com/location',
@@ -50,24 +50,21 @@ class ScanEntityTest extends TestCase
      */
     public function testCall(): void
     {
-        $statusCodes = [
-            200 => true,
-            301 => false,
-            404 => false,
-        ];
-
-        foreach ($statusCodes as $code => $expectedValue) {
-            $this->ScanEntity->set('code', $code);
+        foreach ([
+             200 => true,
+             301 => false,
+             404 => false,
+         ] as $code => $expectedValue) {
+            $this->ScanEntity['code'] = $code;
             $this->assertEquals($expectedValue, $this->ScanEntity->isSuccess());
         }
-        $statusCodes = [
-            200 => false,
-            301 => true,
-            404 => false,
-        ];
 
-        foreach ($statusCodes as $code => $expectedValue) {
-            $this->ScanEntity->set('code', $code);
+        foreach ([
+             200 => false,
+             301 => true,
+             404 => false,
+         ] as $code => $expectedValue) {
+            $this->ScanEntity['code'] = $code;
             $this->assertEquals($expectedValue, $this->ScanEntity->isRedirect());
         }
 
@@ -96,20 +93,10 @@ class ScanEntityTest extends TestCase
      */
     public function testOffsetMethods(): void
     {
-        $this->ScanEntity['name'] = 'value';
-        $this->assertSame('value', $this->ScanEntity['name']);
-        $this->assertTrue(isset($this->ScanEntity['name']));
-        unset($this->ScanEntity['name']);
-        $this->assertFalse(isset($this->ScanEntity['name']));
-    }
-
-    /**
-     * @test
-     * @uses \LinkScanner\ScanEntity::get()
-     * @uses \LinkScanner\ScanEntity::set()
-     */
-    public function testGetAndSetMethods(): void
-    {
-        $this->markTestIncomplete();
+        $this->ScanEntity->key = 'value';
+        $this->assertSame('value', $this->ScanEntity->key);
+        $this->assertTrue(isset($this->ScanEntity->key));
+        unset($this->ScanEntity->key);
+        $this->assertFalse(isset($this->ScanEntity->key));
     }
 }